+++ src/cells/EditorCell.h
@@ -564,7 +564,7 @@ private:
     The current behavior is O(n^2) (scanning the text needs linear time and for each word 
     the indentation algorithm scans the text again) which is unfortunate.
   */
-  void HandleSoftLineBreaks_Code(StyledText *&lastSpace, wxCoord &lineWidth, const wxString &token, unsigned long charInCell,
+  void HandleSoftLineBreaks_Code(StyledText *&lastSpace, wxCoord &lineWidth, const wxString &token, size_t charInCell,
                                  wxString &text, const size_t &lastSpacePos, wxCoord &indentationPixels);
 
   /*! How many chars do we need to indent text at the position the caret is currently at?
